requirements.txt
# Core dependencies
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Computer Vision
opencv-python==4.8.0.76
Pillow==10.0.0

# Data Science
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# XAI Libraries
captum==0.6.0
grad-cam==1.4.8
pytorch-grad-cam==1.4.8
shap==0.42.1

# Utilities
tqdm==4.66.1
jupyter==1.0.0
notebook==7.0.2
ipykernel==6.25.0
kaggle==1.5.16